Rename EventForm to EventRegistrationForm and remove Form from form subdirectories
